wsbsd: Remove unnecessary options to simplify make.conf
The make.conf begins as a combination of /etc/make.conf, /var/db/ports,
/usr/local/etc/ports.conf extracted from the old, broken workstation,
but a lot of options were added from my personal computer shortly after
the conversion to wspkg. Sometimes I even didn't remember what an
option really did and why it was added. Some options were added simply
because it looked cool without doing any testing. Setting more options
in make.conf means it is more likely to see build troubles in poudriere.

To avoid wasting the time of system administrators, options specific to
a port are removed unless it is known to have a use case. Redundant
sets of global options in per-port options are also removed. This
greatly reduces the length of make.conf and hopefully make workstations
easier to maintain.